all: hacks to make sizes consistent for DPI changes
A DPI change happens when a window is moved to a monitor with a
different DPI.

Margins, scrollbar width, and font size should be consistent now
when changing DPIs. Since the scroll point is recorded as a pixel
value, there is still a bug: when the DPI changes, the portion of
the editor that is in view changes. I am planning to refactor all
editor drawing code to keep track of points rather than pixels,
and record changes to pixels-per-point.